Create branch with some potential locking fixes
There are two weird locking scenarios that result in undefined behavior:

1) When creating a thread, we register it via ast_register_thread before
we initialize the mutex in the lock_info stored on the thread's
threadstorage.  However, the lock_info object will be obtained from
threadstorage when the thread registry's RWLOCK is manipulated during
ast_register_thread.  Since we interact with the mutex on the thread
(which will be a de-facto 'normal' pthead mutex) and then re-initialize
it as a recursive mutex, this effectively results in undefined behavior.

2) There is a race condition when creating the recursive mutex that
protects the lock information for a given mutex.  Two simultaneous attempts
to grab that mutex will start off with a NULL lock information object,
resulting in a race to initialize it.  This can result in two initializations
of the same mutex, which is also undefined.

This patch moves some stuff around and puts a global lock around the
recursive lock initialization.  No idea if this will help the goofy
lock freeze ups we see with DEBUG_THREADS enabled, but its a start.