Adding a new structure for sip2cause.
By adding a dynamic structure instead of hardcoding the conversion, we open up
for a new configuration that can change this.

Trying to make up my mind if I just want to add a new dynamic table that is
checked first, or allow editing of this one. 

But first, try to figure out if this works.