Start filling in information on ast_sip_send_request()
I had to modify the method by adding an endpoint because I realized
that for out-of-dialog requests, we did not specify where to send the
request.

The out-of-dialog request-sending method is a stub at the moment since
we haven't defined the ast_sip_endpoint method. Once it's defined, we
can fill in the details and send the request as desired.