Start filling in the on_rx_request method for the session module.
At this point, the most that can happen is that the INVITE will come in
and we'll print a NOTICE level message saying that we got the INVITE.

This can't be tested yet because res_sip does not set up any transports,
meaning that the PJSIP endpoint cannot listen for incoming SIP requests
anywhere.

A good idea would be to set up a hard-coded transport for now until configuration
comes into play.