Fix mishandling of INVITE state during error conditions.
We were attempting to use pjsip_inv_end_session(), which results
in an assertion being raised in PJSIP because the inv_session is
in an invalid state.

What we have to do instead is to use pjsip_inv_initial_answer()
up until we send a 100 Trying (or other provisional response). After
that point, we can end the session by using pjsip_inv_end_session().

This also moves the transmission of the 100 Trying response to incoming
INVITEs from chan_gulp into res_sip_session just before calling out
to supplements.