Write a module that implements RFC3326.
This adds a Reason header to outgoing SIP requests and responses with the Q.850 cause code. For
incoming SIP requests and responses the header is parsed and the cause code in it set on the
channel.

This serves as both a functional implementation and an example of what can be done with the new SIP
architecture.