Backport PJSIP trunk function to clone pjsip_rx_data.
This function is going to be necessary for us to use a threadpool the way we wish.
The function will be present in PJSIP version 2.0.5 and is already in PJSIP trunk.
For now, we are backporting the function where we need it. Whenever PJSIP 2.0.5 is
released, we will  not need to maintain this backport any further.