Add some verification to auth section handling.
This adds some integrity checking to auth handling after the
sorcery objectset has been applied.

In order to do this, I changed the sorcery API's apply handler
callback to return an int so that it is possible to indicate
a failure to apply. This meant having to change the two places
where an apply handler is used to return a failure result
if need be.