First measure in storing authentication challenges instead of being stateless.
We now generate a nonce for each authentication challenge and store the challenge
in a container. The container is keyed off the nonce. This way, when a request comes
in, we can first find if we have a challenge with that nonce and then move on to
verification.

What needs to be added now is a destruction mechanism for authentication challenges.
This includes destroying challenges as requests with credentials arrive as well as
timing out authentication challenges that receive no answers.

This code is untested. It will be tested once further improvements are made.