Completely change the scheme for verifying authentication credentials.
Due to the pure stateless manner in which things had been done, Joshua
had pointed out that we were susceptible to replay attacks. My initial
reaction to this was to send a randomized nonce in challenges, and then
save this value in a container.

This approach presents several issues:
1) Having stuff in a container means there is a contention point between
SIP threads. While testing has not shown this to be a problem, it's easy
to see how this might be a performance bottleneck.
2) There is a tremendous amount of overhead involved in making sure to
expire challenges within a certain amount of time.

After a conversation with Joshua, we now have a new approach.

Now we generate a nonce that is based on a hash of the timestamp, source
IP address and port, an entity ID, and the realm. When we receive an
INVITE with authentication info, we can calculate the nonce a second
time and compare with the nonce they provide in their credentials
to be sure it seems sane.